backend-api/requirements.txt
fastapi==0.115.5
gunicorn==21.2.0
uvicorn[standard]==0.32.1

pydantic==2.10.3
pydantic-settings==2.10.1

python-multipart
python-dotenv
aiofiles==24.1.0

authlib
python-jose
PyJWT
itsdangerous
email-validator
httpx

pymongo
motor

passlib[bcrypt]==1.7.4
bcrypt==4.1.2

cloudinary

# Basic image handling (no ML)
pillow==11.0.0
